Fix symbol demangling on FreeBSD
To produce meaningful stack traces for native code, the JVM will try to
demangle symbols before presenting them. This makes C++ function names
appear as they do in the code with namespace and class names represented
properly.

To do this (on BSD's) it calls the system provided abi::__cxa_demangle
function. At least the FreeBSD implementation of this function will
return with a valid string containing garbage (but consistent carbage),
and a status of 0, if the origina symbol was _not_ mangled.

This patch adds a quick filter to ensure that only mangled symbols are
passed to the function. This ensures that the original symbol is
preserved in these cases.

This work was sponsored by: The FreeBSD Foundation